Rellocating heredoc functions into ft_heredoc.c
